Add imgui node editor implementation
Fix use of multiple render window, and allow user to create it at runtime

+++ VulkanDecouverte/Mesh.h
@@ -38,7 +38,7 @@ struct Vertex
 struct MeshData
 {
     std::vector<Vertex> vertices;
-    std::vector<uint16_t> indices;
+    std::vector<uint32_t> indices;
 
     static MeshData& Cube()
     {
